enable warning check and fix warnings. (#760)
* enable warning check and fix warnings.

* sync and merge

* disable warning 4146

* fix build issue.

* fix warnings

+++ onnx/defs/shape_inference.h
@@ -33,7 +33,7 @@ inline bool getRepeatedAttribute(InferenceContext& ctx,
 }
 
 inline bool hasExactlyNInputTypes(InferenceContext& ctx, int n, const std::string& opname) {
-  if (ctx.getNumInputs() != n) {
+  if (static_cast<int>(ctx.getNumInputs()) != n) {
     throw std::runtime_error(opname + " has wrong number of inputs");
   }
   for (int i = 0; i < n; i++) {
@@ -50,7 +50,7 @@ inline void propagateElemTypeFromInputToOutput(InferenceContext& ctx, size_t inp
 
 inline void appendSingleDimCopiedFromInputTypeToOutputType(InferenceContext& ctx, size_t inputIndex, size_t outputIndex, size_t fromDimIndex) {
   auto* dim = ctx.getOutputType(outputIndex)->mutable_shape()->add_dim();
-  *dim = ctx.getInputType(inputIndex)->shape().dim(fromDimIndex);
+  *dim = ctx.getInputType(inputIndex)->shape().dim(static_cast<int>(fromDimIndex));
 }
 
 inline void propagateShapeFromInputToOutput(InferenceContext& ctx, size_t inputIndex, size_t outputIndex) {
